backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
python-decouple
Pillow
openai
django-filter
google-cloud-aiplatform
